JP-BE-PY-batch-1/flask-app/cryptography/requirements.txt
pycryptodome
cryptography
bcrypt
rsa
pyjwt